feat(domains): per-domain PHP Settings tab on the Web Domain page
GH #1543: give the tenant Web Domain page a PHP Settings tab holding the
DOMAIN-scoped PHP controls — this domain's PHP version (POST/DELETE
/domains/:id/php-pool) plus its php.ini limit overrides (GET/PATCH
/domains/:id/php-settings) and its environment variables
(/domains/:id/env-vars).

The account / pool-level PHP tabs (CLI & Composer, Performance, OPcache,
Extensions, Xdebug) are all /me/... and per-version-pool — a pool is
shared by every domain running that PHP version, so they are not
per-domain and cannot live on a single-domain page. They stay on the
standalone PHP Settings page, which therefore stays. (Whether that page's
domain-selector "Version & Domains" tab is still worth keeping now every
domain page has its own PHP Settings tab is a separate decision, like the
Logs nav-rename in the previous slice.)

Extraction, not duplication (the A3/A5 pattern): the per-domain version +
limits form is lifted into a new audience-neutral
components/domains/DomainPHPSettingsPanel.tsx ({ domainId }); the
standalone page's "Version & Domains" tab now renders it behind its domain
picker, so there is one copy of the form, the option lists and the save
logic. DomainEnvVarsCard gains an optional domainId prop that binds the
domain and hides its own picker; the Web Domain tab passes it, the
standalone Environment tab keeps calling it bare.

Latent-bug fix carried in the extraction: the load effect now
resetFields() BEFORE setFieldsValue(), because setFieldsValue does not
clear touched state — without it, switching the selected domain on the
standalone page left the Save button enabled on a stale dirty flag. Unit
tests pin the domain binding at both new call sites (the panel DELETEs the
given domain's pool; the env card loads the given domain and never lists
/domains).

The tab sits after DNS and before Redirects; it is not cap-gated (the
standalone selector already lists every domain).

GH #1543
